feat(signers-page): added signers chart + new layout
