+++ src/app/signers/consts.ts
@@ -1,4 +1,4 @@
-export const SIGNER_KEY_MAP = {
+export const SIGNER_KEY_MAP: Record<string, { poolOperator: string; name: string }> = {
   '03e0df37e83e43847625a0320456cb9758050a61ce76c2c130bf50242f27ba6d54': {
     poolOperator: 'SP4SZE494VC2YC5JYG7AYFQ44F5Q4PYV7DVMDPBG.stacking-pool-signer-alum-labs-v1',
     name: 'Alum Labs',
@@ -64,3 +64,5 @@ export const SIGNER_KEY_MAP = {
     name: 'Luxor',
   },
 };
+
+
